requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-telegram-bot==20.7
gspread==6.0.2
google-auth==2.27.0
google-auth-httplib2==0.2.0
requests==2.31.0
APScheduler==3.10.4
psycopg2-binary==2.9.9
